Add no inhibitor combat strategy (#39)
* add no inhibitor combat strategy

* add sigmoid card sim output

* add no inhibitor card sim output

* use sigmoid in no inhibitor

* add baseline

* format

* Delete card-sim.md (#40)

* put sigmoid strategy back
