+++ kaori/plugins/gacha/engine/__init__.py
@@ -61,10 +61,11 @@ cs_sigmoid_v1 = SigmoidV1(rarities=rarities,
                           crit_multiplier=CRIT_MULTIPLIER,
                           stat_curvatures=stat_curvatures)
 
-cs_rid_v1 = RidV1(rarities=rarities,
+cs_no_inhibitor_v0 = NoInhibitorV0(rarities=rarities,
                   stats=stats,
                   natures=natures,
-                  crit_multiplier=CRIT_MULTIPLIER)
+                  crit_multiplier=CRIT_MULTIPLIER,
+                  stat_curvatures=stat_curvatures)
 
 combat_strat = cs_sigmoid_v1
 user_combat_strat = os.environ.get('COMBAT_STRATEGY')
