+++ kaori/plugins/gacha/engine/core/combat_strategies/__init__.py
@@ -1,3 +1,3 @@
 from .abc import CombatStrategyABC
-from .rid import RidV1
+from .no_inhibitor import NoInhibitorV0
 from .sigmoid import SigmoidV1
